+++ CHANGELOG.md
@@ -25,6 +25,10 @@
 
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- The compiler now correctly prints types, including module names, when
+  reporting diagnostics for inexhaustive case expressions.
+  ([Surya Rose](https://github.com/gearsdatapacks))
+
 ### Formatter
 
 ### Language Server